refactor document loader to a global util
