+++ app/features/rubric_generator/core.py
@@ -1,7 +1,7 @@
 from app.services.logger import setup_logger
 from app.api.error_utilities import LoaderError, ToolExecutorError
 from app.services.schemas import RubricGeneratorArgs
-from app.features.rubric_generator.document_loaders import get_docs
+from app.utils.document_loaders import get_docs
 from app.features.rubric_generator.tools import RubricGenerator
 
 logger = setup_logger()
